+++ src/base/Element.php
@@ -491,7 +491,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public static function refHandle()
+    public static function refHandle(): ?string
     {
         return null;
     }
@@ -567,7 +567,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public static function findOne($criteria = null)
+    public static function findOne($criteria = null): ?ElementInterface
     {
         return static::findByCondition($criteria, true);
     }
@@ -2083,7 +2083,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function getId()
+    public function getId(): ?int
     {
         return $this->id;
     }
@@ -2267,7 +2267,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function getFieldLayout()
+    public function getFieldLayout(): ?FieldLayout
     {
         if ($this->fieldLayoutId) {
             return Craft::$app->getFields()->getLayoutById($this->fieldLayoutId);
@@ -2300,7 +2300,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function getUriFormat()
+    public function getUriFormat(): ?string
     {
         return null;
     }
@@ -2376,7 +2376,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function getUrl()
+    public function getUrl(): ?string
     {
         if ($this->uri === null) {
             return null;
@@ -2389,7 +2389,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function getLink()
+    public function getLink(): ?Markup
     {
         if (($url = $this->getUrl()) === null) {
             return null;
@@ -2429,7 +2429,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function getRef()
+    public function getRef(): ?string
     {
         return null;
     }
@@ -2484,7 +2484,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function getCpEditUrl()
+    public function getCpEditUrl(): ?string
     {
         $cpEditUrl = $this->cpEditUrl();
 
@@ -2581,7 +2581,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function getThumbUrl(int $size)
+    public function getThumbUrl(int $size): ?string
     {
         return null;
     }
@@ -2605,7 +2605,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function getEnabledForSite(int $siteId = null)
+    public function getEnabledForSite(int $siteId = null): ?bool
     {
         if ($siteId === null) {
             $siteId = $this->siteId;
@@ -2622,7 +2622,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function setEnabledForSite($enabledForSite)
+    public function setEnabledForSite($enabledForSite): void
     {
         if (is_array($enabledForSite)) {
             foreach ($enabledForSite as &$value) {
@@ -2637,7 +2637,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function getStatus()
+    public function getStatus(): ?string
     {
         if ($this->archived) {
             return self::STATUS_ARCHIVED;
@@ -2673,7 +2673,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function getNext($criteria = false)
+    public function getNext($criteria = false): ?ElementInterface
     {
         if ($criteria !== false || $this->_nextElement === null) {
             return $this->_getRelativeElement($criteria, 1);
@@ -2689,7 +2689,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function getPrev($criteria = false)
+    public function getPrev($criteria = false): ?ElementInterface
     {
         if ($criteria !== false || $this->_prevElement === null) {
             return $this->_getRelativeElement($criteria, -1);
@@ -2705,7 +2705,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function setNext($element)
+    public function setNext($element): void
     {
         $this->_nextElement = $element;
     }
@@ -2713,7 +2713,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function setPrev($element)
+    public function setPrev($element): void
     {
         $this->_prevElement = $element;
     }
@@ -2721,7 +2721,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function getParent()
+    public function getParent(): ?ElementInterface
     {
         if ($this->_parent === null) {
             $ancestors = $this->getAncestors(1);
@@ -2755,7 +2755,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function setParent(ElementInterface $parent = null)
+    public function setParent(ElementInterface $parent = null): void
     {
         $this->_parent = $parent;
 
@@ -2837,7 +2837,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function getPrevSibling()
+    public function getPrevSibling(): ?ElementInterface
     {
         if ($this->_prevSibling === null) {
             /** @var ElementQuery $query */
@@ -2859,7 +2859,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function getNextSibling()
+    public function getNextSibling(): ?ElementInterface
     {
         if ($this->_nextSibling === null) {
             /** @var ElementQuery $query */
@@ -2975,7 +2975,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function offsetExists($offset)
+    public function offsetExists($offset): bool
     {
         return $offset === 'title' || $this->hasEagerLoadedElements($offset) || parent::offsetExists($offset) || $this->fieldByHandle($offset);
     }
@@ -2983,7 +2983,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    function getAttributeStatus(string $attribute)
+    function getAttributeStatus(string $attribute): ?array
     {
         if ($this->isAttributeModified($attribute)) {
             return [
@@ -3109,7 +3109,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function setDirtyAttributes(array $names, bool $merge = true)
+    public function setDirtyAttributes(array $names, bool $merge = true): void
     {
         if ($merge) {
             $this->_dirtyAttributes = array_merge($this->_dirtyAttributes, array_flip($names));
@@ -3129,7 +3129,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function getTitleTranslationDescription()
+    public function getTitleTranslationDescription(): ?string
     {
         return ElementHelper::translationDescription(Field::TRANSLATION_METHOD_SITE);
     }
@@ -3178,7 +3178,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function setFieldValues(array $values)
+    public function setFieldValues(array $values): void
     {
         foreach ($values as $fieldHandle => $value) {
             $this->setFieldValue($fieldHandle, $value);
@@ -3199,7 +3199,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function setFieldValue(string $fieldHandle, $value)
+    public function setFieldValue(string $fieldHandle, $value): void
     {
         $behavior = $this->getBehavior('customFields');
         $behavior->$fieldHandle = $value;
@@ -3351,7 +3351,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function markAsDirty()
+    public function markAsDirty(): void
     {
         $this->_allDirty = true;
     }
@@ -3359,7 +3359,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function markAsClean()
+    public function markAsClean(): void
     {
         $this->_allDirty = false;
         $this->_dirtyAttributes = [];
@@ -3372,7 +3372,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function setFieldValuesFromRequest(string $paramNamespace = '')
+    public function setFieldValuesFromRequest(string $paramNamespace = ''): void
     {
         $this->setFieldParamNamespace($paramNamespace);
         $values = Craft::$app->getRequest()->getBodyParam($paramNamespace, []);
@@ -3398,7 +3398,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function getFieldParamNamespace()
+    public function getFieldParamNamespace(): ?string
     {
         return $this->_fieldParamNamePrefix;
     }
@@ -3406,7 +3406,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function setFieldParamNamespace(string $namespace)
+    public function setFieldParamNamespace(string $namespace): void
     {
         $this->_fieldParamNamePrefix = $namespace;
     }
@@ -3446,7 +3446,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function getEagerLoadedElements(string $handle)
+    public function getEagerLoadedElements(string $handle): ?array
     {
         if (!isset($this->_eagerLoadedElements[$handle])) {
             return null;
@@ -3461,7 +3461,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function setEagerLoadedElements(string $handle, array $elements)
+    public function setEagerLoadedElements(string $handle, array $elements): void
     {
         switch ($handle) {
             case 'parent':
@@ -3507,7 +3507,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function setEagerLoadedElementCount(string $handle, int $count)
+    public function setEagerLoadedElementCount(string $handle, int $count): void
     {
         $this->_eagerLoadedElementCounts[$handle] = $count;
     }
@@ -3523,7 +3523,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function setRevisionCreatorId(int $creatorId = null)
+    public function setRevisionCreatorId(int $creatorId = null): void
     {
         $this->revisionCreatorId = $creatorId;
     }
@@ -3531,7 +3531,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function setRevisionNotes(string $notes = null)
+    public function setRevisionNotes(string $notes = null): void
     {
         $this->revisionNotes = $notes;
     }
@@ -3539,7 +3539,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function getCurrentRevision()
+    public function getCurrentRevision(): ?ElementInterface
     {
         if (!$this->id) {
             return null;
@@ -3943,7 +3943,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function afterSave(bool $isNew)
+    public function afterSave(bool $isNew): void
     {
         // Tell the fields about it
         foreach ($this->fieldLayoutFields() as $field) {
@@ -3961,7 +3961,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function afterPropagate(bool $isNew)
+    public function afterPropagate(bool $isNew): void
     {
         // Tell the fields about it
         foreach ($this->fieldLayoutFields() as $field) {
@@ -3998,7 +3998,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function afterDelete()
+    public function afterDelete(): void
     {
         // Tell the fields about it
         foreach ($this->fieldLayoutFields() as $field) {
@@ -4033,7 +4033,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function afterRestore()
+    public function afterRestore(): void
     {
         // Tell the fields about it
         foreach ($this->fieldLayoutFields() as $field) {
@@ -4063,7 +4063,7 @@ abstract class Element extends Component implements ElementInterface
     /**
      * @inheritdoc
      */
-    public function afterMoveInStructure(int $structureId)
+    public function afterMoveInStructure(int $structureId): void
     {
         // Trigger an 'afterMoveInStructure' event
         if ($this->hasEventHandlers(self::EVENT_AFTER_MOVE_IN_STRUCTURE)) {