+++ src/base/ElementInterface.php
@@ -54,7 +54,7 @@ interface ElementInterface extends ComponentInterface
      *
      * @return string|null The reference handle, or null if the element type doesn’t support reference tags
      */
-    public static function refHandle();
+    public static function refHandle(): ?string;
 
     /**
      * Returns whether Craft should keep track of attribute and custom field changes made to this element type,
@@ -186,7 +186,7 @@ interface ElementInterface extends ComponentInterface
      * @param mixed $criteria The element ID or a set of element criteria parameters
      * @return static|null Element instance matching the condition, or null if nothing matches.
      */
-    public static function findOne($criteria = null);
+    public static function findOne($criteria = null): ?ElementInterface;
 
     /**
      * Returns a list of elements that match the specified ID(s) or a set of element criteria parameters.
@@ -521,7 +521,7 @@ interface ElementInterface extends ComponentInterface
      * @internal This method is required by [[\yii\web\IdentityInterface]], but might as well
      * go here rather than only in [[\craft\elements\User]].
      */
-    public function getId();
+    public function getId(): ?int;
 
     /**
      * Returns whether this is a draft.
@@ -624,7 +624,7 @@ interface ElementInterface extends ComponentInterface
      *
      * @return FieldLayout|null
      */
-    public function getFieldLayout();
+    public function getFieldLayout(): ?FieldLayout;
 
     /**
      * Returns the site the element is associated with.
@@ -665,7 +665,7 @@ interface ElementInterface extends ComponentInterface
      * @see hasUris()
      * @see getRoute()
      */
-    public function getUriFormat();
+    public function getUriFormat(): ?string;
 
     /**
      * Returns the search keywords for a given search attribute.
@@ -700,14 +700,14 @@ interface ElementInterface extends ComponentInterface
      *
      * @return string|null
      */
-    public function getUrl();
+    public function getUrl(): ?string;
 
     /**
      * Returns an anchor pre-filled with this element’s URL and title.
      *
      * @return Markup|null
      */
-    public function getLink();
+    public function getLink(): ?Markup;
 
     /**
      * Returns what the element should be called within the control panel.
@@ -731,7 +731,7 @@ interface ElementInterface extends ComponentInterface
      *
      * @return string|null
      */
-    public function getRef();
+    public function getRef(): ?string;
 
     /**
      * Returns whether the current user can edit the element.
@@ -753,7 +753,7 @@ interface ElementInterface extends ComponentInterface
      *
      * @return string|null
      */
-    public function getCpEditUrl();
+    public function getCpEditUrl(): ?string;
 
     /**
      * Returns the additional locations that should be available for previewing the element, besides its primary [[getUrl()|URL]].
@@ -780,7 +780,7 @@ interface ElementInterface extends ComponentInterface
      * @param int $size The maximum width and height the thumbnail should have.
      * @return string|null
      */
-    public function getThumbUrl(int $size);
+    public function getThumbUrl(int $size): ?string;
 
     /**
      * Returns whether the element’s thumbnail should have a checkered background.
@@ -808,7 +808,7 @@ interface ElementInterface extends ComponentInterface
      * passed, but that site’s status wasn’t provided via [[setEnabledForSite()]].
      * @since 3.4.0
      */
-    public function getEnabledForSite(int $siteId = null);
+    public function getEnabledForSite(int $siteId = null): ?bool;
 
     /**
      * Sets whether the element is enabled for the current site.
@@ -816,16 +816,17 @@ interface ElementInterface extends ComponentInterface
      * This can also be set to an array of site ID/site-enabled mappings.
      *
      * @param bool|bool[] $enabledForSite
+     * @return void
      * @since 3.4.0
      */
-    public function setEnabledForSite($enabledForSite);
+    public function setEnabledForSite($enabledForSite): void;
 
     /**
      * Returns the element’s status.
      *
      * @return string|null
      */
-    public function getStatus();
+    public function getStatus(): ?string;
 
     /**
      * Returns the same element in other locales.
@@ -840,7 +841,7 @@ interface ElementInterface extends ComponentInterface
      * @param mixed $criteria
      * @return static|null
      */
-    public function getNext($criteria = false);
+    public function getNext($criteria = false): ?ElementInterface;
 
     /**
      * Returns the previous element relative to this one, from a given set of criteria.
@@ -848,14 +849,15 @@ interface ElementInterface extends ComponentInterface
      * @param mixed $criteria
      * @return static|null
      */
-    public function getPrev($criteria = false);
+    public function getPrev($criteria = false): ?ElementInterface;
 
     /**
      * Sets the default next element.
      *
      * @param static|false $element
+     * @return void
      */
-    public function setNext($element);
+    public function setNext($element): void;
 
     /**
      * Sets the default previous element.
@@ -863,14 +865,14 @@ interface ElementInterface extends ComponentInterface
      * @param static|false $element
      * @return void
      */
-    public function setPrev($element);
+    public function setPrev($element): void;
 
     /**
      * Returns the element’s parent.
      *
      * @return static|null
      */
-    public function getParent();
+    public function getParent(): ?ElementInterface;
 
     /**
      * Returns the parent element’s URI, if there is one.
@@ -885,8 +887,9 @@ interface ElementInterface extends ComponentInterface
      * Sets the element’s parent.
      *
      * @param static|null $parent
+     * @return void
      */
-    public function setParent(ElementInterface $parent = null);
+    public function setParent(ElementInterface $parent = null): void;
 
     /**
      * Returns the element’s ancestors.
@@ -923,14 +926,14 @@ interface ElementInterface extends ComponentInterface
      *
      * @return static|null
      */
-    public function getPrevSibling();
+    public function getPrevSibling(): ?ElementInterface;
 
     /**
      * Returns the element’s next sibling.
      *
      * @return static|null
      */
-    public function getNextSibling();
+    public function getNextSibling(): ?ElementInterface;
 
     /**
      * Returns whether the element has descendants.
@@ -1008,7 +1011,7 @@ interface ElementInterface extends ComponentInterface
      * @param string|int $offset
      * @return bool
      */
-    public function offsetExists($offset);
+    public function offsetExists($offset): bool;
 
     /**
      * Returns the status of a given attribute.
@@ -1017,7 +1020,7 @@ interface ElementInterface extends ComponentInterface
      * @return array|null
      * @since 3.4.0
      */
-    public function getAttributeStatus(string $attribute);
+    public function getAttributeStatus(string $attribute): ?array;
 
     /**
      * Returns the attribute names that have been updated on the canonical element since the last time it was
@@ -1076,10 +1079,11 @@ interface ElementInterface extends ComponentInterface
      *
      * @param string[] $names
      * @param bool $merge Whether these attributes should be merged with existing dirty attributes
+     * @return void
      * @see getDirtyAttributes()
      * @since 3.5.0
      */
-    public function setDirtyAttributes(array $names, bool $merge = true);
+    public function setDirtyAttributes(array $names, bool $merge = true): void;
 
     /**
      * Returns whether the Title field should be shown as translatable in the UI.
@@ -1098,7 +1102,7 @@ interface ElementInterface extends ComponentInterface
      * @return string|null
      * @since 3.5.0
      */
-    public function getTitleTranslationDescription();
+    public function getTitleTranslationDescription(): ?string;
 
     /**
      * Returns the Title’s translation key.
@@ -1144,8 +1148,9 @@ interface ElementInterface extends ComponentInterface
      * Sets the element’s custom field values.
      *
      * @param array $values The custom field values (handle => value)
+     * @return void
      */
-    public function setFieldValues(array $values);
+    public function setFieldValues(array $values): void;
 
     /**
      * Returns the value for a given field.
@@ -1161,8 +1166,9 @@ interface ElementInterface extends ComponentInterface
      *
      * @param string $fieldHandle The field handle whose value needs to be set
      * @param mixed $value The value to set on the field
+     * @return void
      */
-    public function setFieldValue(string $fieldHandle, $value);
+    public function setFieldValue(string $fieldHandle, $value): void;
 
     /**
      * Returns the field handles that have been updated on the canonical element since the last time it was
@@ -1219,16 +1225,18 @@ interface ElementInterface extends ComponentInterface
     /**
      * Marks all fields and attributes as dirty.
      *
+     * @return void
      * @since 3.4.10
      */
-    public function markAsDirty();
+    public function markAsDirty(): void;
 
     /**
      * Resets the record of dirty attributes and fields.
      *
+     * @return void
      * @since 3.4.0
      */
-    public function markAsClean();
+    public function markAsClean(): void;
 
     /**
      * Returns the cache tags that should be cleared when this element is saved.
@@ -1242,22 +1250,24 @@ interface ElementInterface extends ComponentInterface
      * Sets the element’s custom field values, when the values have come from post data.
      *
      * @param string $paramNamespace The field param namespace
+     * @return void
      */
-    public function setFieldValuesFromRequest(string $paramNamespace);
+    public function setFieldValuesFromRequest(string $paramNamespace): void;
 
     /**
      * Returns the namespace used by custom field params on the request.
      *
      * @return string|null The field param namespace
      */
-    public function getFieldParamNamespace();
+    public function getFieldParamNamespace(): ?string;
 
     /**
      * Sets the namespace used by custom field params on the request.
      *
      * @param string $namespace The field param namespace
+     * @return void
      */
-    public function setFieldParamNamespace(string $namespace);
+    public function setFieldParamNamespace(string $namespace): void;
 
     /**
      * Returns the name of the table this element’s content is stored in.
@@ -1294,15 +1304,16 @@ interface ElementInterface extends ComponentInterface
      * @param string $handle The handle of the eager-loaded elements
      * @return ElementInterface[]|null The eager-loaded elements, or null if they hadn't been eager-loaded
      */
-    public function getEagerLoadedElements(string $handle);
+    public function getEagerLoadedElements(string $handle): ?array;
 
     /**
      * Sets some eager-loaded elements on a given handle.
      *
      * @param string $handle The handle that was used to eager-load the elements
      * @param ElementInterface[] $elements The eager-loaded elements
+     * @return void
      */
-    public function setEagerLoadedElements(string $handle, array $elements);
+    public function setEagerLoadedElements(string $handle, array $elements): void;
 
     /**
      * Returns the count of eager-loaded elements for a given handle.
@@ -1318,9 +1329,10 @@ interface ElementInterface extends ComponentInterface
      *
      * @param string $handle The handle to load the elements with in the future
      * @param int $count The eager-loaded element count
+     * @return void
      * @since 3.4.0
      */
-    public function setEagerLoadedElementCount(string $handle, int $count);
+    public function setEagerLoadedElementCount(string $handle, int $count): void;
 
     /**
      * Returns whether the element’s content is "fresh" (unsaved and without validation errors).
@@ -1333,17 +1345,19 @@ interface ElementInterface extends ComponentInterface
      * Sets the revision creator ID to be saved.
      *
      * @param int|null $creatorId
+     * @return void
      * @since 3.2.0
      */
-    public function setRevisionCreatorId(int $creatorId = null);
+    public function setRevisionCreatorId(int $creatorId = null): void;
 
     /**
      * Sets the revision notes to be saved.
      *
      * @param string|null $notes
+     * @return void
      * @since 3.2.0
      */
-    public function setRevisionNotes(string $notes = null);
+    public function setRevisionNotes(string $notes = null): void;
 
     /**
      * Returns the element’s current revision, if one exists.
@@ -1351,7 +1365,7 @@ interface ElementInterface extends ComponentInterface
      * @return static|null
      * @since 3.2.0
      */
-    public function getCurrentRevision();
+    public function getCurrentRevision(): ?ElementInterface;
 
     // Indexes, etc.
     // -------------------------------------------------------------------------
@@ -1430,8 +1444,9 @@ interface ElementInterface extends ComponentInterface
      * Performs actions after an element is saved.
      *
      * @param bool $isNew Whether the element is brand new
+     * @return void
      */
-    public function afterSave(bool $isNew);
+    public function afterSave(bool $isNew): void;
 
     /**
      * Performs actions after an element is fully saved and propagated to other sites.
@@ -1441,9 +1456,10 @@ interface ElementInterface extends ComponentInterface
      * :::
      *
      * @param bool $isNew Whether the element is brand new
+     * @return void
      * @since 3.2.0
      */
-    public function afterPropagate(bool $isNew);
+    public function afterPropagate(bool $isNew): void;
 
     /**
      * Performs actions before an element is deleted.
@@ -1454,8 +1470,10 @@ interface ElementInterface extends ComponentInterface
 
     /**
      * Performs actions after an element is deleted.
+     *
+     * @return void
      */
-    public function afterDelete();
+    public function afterDelete(): void;
 
     /**
      * Performs actions before an element is restored.
@@ -1468,9 +1486,10 @@ interface ElementInterface extends ComponentInterface
     /**
      * Performs actions after an element is restored.
      *
+     * @return void
      * @since 3.1.0
      */
-    public function afterRestore();
+    public function afterRestore(): void;
 
     /**
      * Performs actions before an element is moved within a structure.
@@ -1484,6 +1503,7 @@ interface ElementInterface extends ComponentInterface
      * Performs actions after an element is moved within a structure.
      *
      * @param int $structureId The structure ID
+     * @return void
      */
-    public function afterMoveInStructure(int $structureId);
+    public function afterMoveInStructure(int $structureId): void;
 }