+++ src/elements/Entry.php
@@ -39,6 +39,7 @@ use craft\helpers\Html;
 use craft\helpers\Json;
 use craft\helpers\UrlHelper;
 use craft\models\EntryType;
+use craft\models\FieldLayout;
 use craft\models\Section;
 use craft\models\Site;
 use craft\records\Entry as EntryRecord;
@@ -106,7 +107,7 @@ class Entry extends Element
     /**
      * @inheritdoc
      */
-    public static function refHandle()
+    public static function refHandle(): ?string
     {
         return 'entry';
     }
@@ -931,7 +932,7 @@ class Entry extends Element
      * @inheritdoc
      * @throws InvalidConfigException if [[siteId]] is not set to a site ID that the entry's section is enabled for
      */
-    public function getUriFormat()
+    public function getUriFormat(): ?string
     {
         $sectionSiteSettings = $this->getSection()->getSiteSettings();
 
@@ -998,7 +999,7 @@ class Entry extends Element
      *
      * @return string|null
      */
-    public function getRef()
+    public function getRef(): ?string
     {
         return $this->getSection()->handle . '/' . $this->slug;
     }
@@ -1014,7 +1015,7 @@ class Entry extends Element
     /**
      * @inheritdoc
      */
-    public function getTitleTranslationDescription()
+    public function getTitleTranslationDescription(): ?string
     {
         return ElementHelper::translationDescription($this->getType()->titleTranslationMethod);
     }
@@ -1031,7 +1032,7 @@ class Entry extends Element
     /**
      * @inheritdoc
      */
-    public function getFieldLayout()
+    public function getFieldLayout(): ?FieldLayout
     {
         if (($fieldLayout = parent::getFieldLayout()) !== null) {
             return $fieldLayout;
@@ -1176,7 +1177,7 @@ class Entry extends Element
     /**
      * @inheritdoc
      */
-    public function getStatus()
+    public function getStatus(): ?string
     {
         $status = parent::getStatus();
 
@@ -1303,7 +1304,7 @@ class Entry extends Element
     /**
      * @inheritdoc
      */
-    public function setEagerLoadedElements(string $handle, array $elements)
+    public function setEagerLoadedElements(string $handle, array $elements): void
     {
         if ($handle === 'author') {
             $this->_author = $elements[0] ?? false;
@@ -1584,7 +1585,7 @@ EOD;
      * @inheritdoc
      * @throws Exception if reasons
      */
-    public function afterSave(bool $isNew)
+    public function afterSave(bool $isNew): void
     {
         if (!$this->propagating) {
             $section = $this->getSection();
@@ -1646,7 +1647,7 @@ EOD;
     /**
      * @inheritdoc
      */
-    public function afterPropagate(bool $isNew)
+    public function afterPropagate(bool $isNew): void
     {
         parent::afterPropagate($isNew);
 
@@ -1691,7 +1692,7 @@ EOD;
     /**
      * @inheritdoc
      */
-    public function afterRestore()
+    public function afterRestore(): void
     {
         $section = $this->getSection();
         if ($section->type === Section::TYPE_STRUCTURE) {
@@ -1715,7 +1716,7 @@ EOD;
     /**
      * @inheritdoc
      */
-    public function afterMoveInStructure(int $structureId)
+    public function afterMoveInStructure(int $structureId): void
     {
         // Was the entry moved within its section's structure?
         $section = $this->getSection();